Create invidual TaskBox component
This component was created to make it easier to add each individual
action (such as delete, edit, copy, etc) to the task list.
